Address bounded run contract review findings
